+++ CHANGES.md
@@ -8,6 +8,7 @@ Requires Menhir 20211230 and OCaml 4.13 or above.
 - API:
   - New `Setup.init` takes a `?versions` dictionary to declare versions
     of external components
+  - New `Utils.parse_version`
 
 # v2.0.6 (December 2024)